requirements.txt
tensorflow
keras
numpy
pillow
fastapi
python-multipart
fastapi[all]